library_manager/requirements.txt
Django==4.2.7
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==20.1.0
psycopg2==2.9.1
django-braces==1.15.0